Conflicts:
	TFC API/TFC/API/Constant/TFCItemID.java
	TFC_Shared/src/TFC/Blocks/Flora/BlockFruitWood.java
	TFC_Shared/src/TFC/Handlers/CraftingHandler.java
	TFC_Shared/src/TFC/TFCItems.java
	api/TFC/API/TFC_ItemHeat.java
	src/main/java/TFC/Blocks/BlockSlab.java
	src/main/java/TFC/Blocks/Devices/BlockForge.java
	src/main/java/TFC/Core/ItemHeat.java
	src/main/java/TFC/Items/ItemIngot.java
	src/main/java/TFC/Items/ItemTerra.java
	src/main/java/TFC/Items/Tools/ItemCustomBucketMilk.java
	src/main/java/TFC/TileEntities/TEBlastFurnace.java
	src/main/java/TFC/TileEntities/TECrucible.java
	src/main/java/TFC/TileEntities/TileEntityAnvil.java
	src/main/java/TFC/TileEntities/TileEntityFireEntity.java
	src/main/java/TFC/TileEntities/TileEntityFirepit.java
	src/main/java/TFC/TileEntities/TileEntityForge.java

+++ src/main/java/TFC/Handlers/CraftingHandler.java
@@ -160,7 +160,7 @@ public class CraftingHandler
 				else if(!e.player.inventory.addItemStackToInventory(new ItemStack(TFCItems.CeramicMold, 1, 1)))
 					e.player.entityDropItem(new ItemStack(TFCItems.CeramicMold, 1, 1), 1);
 
-				short temperature = 0;
+				float temperature = 0;
 				for(int i = 0; i < iinventory.getSizeInventory(); i++)
 				{
 					if(iinventory.getStackInSlot(i) == null)
@@ -172,7 +172,7 @@ public class CraftingHandler
 			}
 			else if(item instanceof ItemMeltedMetal)
 			{
-				short temperature = 0;
+				float temperature = 0;
 				for(int i = 0; i < iinventory.getSizeInventory(); i++)
 				{
 					if(iinventory.getStackInSlot(i) == null)
